+++ src/ai_protocol_mock/engine/__init__.py
@@ -3,13 +3,16 @@
 from ai_protocol_mock.engine.errors import apply_mock_error, is_stream_interrupt
 from ai_protocol_mock.engine.generator import ChatMockContext, generate_chat_response, parse_chat_context
 from ai_protocol_mock.engine.resolver import ContractResolver, ProviderResolution
+from ai_protocol_mock.engine.stream import StreamEncoder, resolve_request_path
 
 __all__ = [
     "ChatMockContext",
     "ContractResolver",
     "ProviderResolution",
+    "StreamEncoder",
     "apply_mock_error",
     "generate_chat_response",
     "is_stream_interrupt",
     "parse_chat_context",
+    "resolve_request_path",
 ]